separate classes Clipboard and ClipboardService
